Fix issue of vanishing diagnostics
